Delete newly created files in Compile method to prevent "File already exists" exception in future file creations with the same name (Closes #399)

+++ Open Judge System/Workers/OJS.Workers.ExecutionStrategies/ExecutionStrategy.cs
@@ -113,7 +113,11 @@
 
             ICompiler compiler = Compiler.CreateCompiler(compilerType);
             var compilerResult = compiler.Compile(compilerPath, submissionFilePath, compilerArguments);
-            File.Delete(submissionFilePath);
+
+            if (File.Exists(submissionFilePath))
+            {
+                File.Delete(submissionFilePath);
+            }
 
             return compilerResult;
         }
